requirements.txt
django==3.1.5
djangorestframework==3.12.2
sqlparse==0.4.1
asgiref==3.3.1
pytz==2020.5
mysqlclient==2.0.3
requests==2.25.1
pytest-django==4.1.0
djangorestframework-simplejwt==4.6.0